Wrap the `app` into a fixture
